Minor User and Role Validation Fix

+++ src/main/java/com/alkemy/ong/services/RoleService.java
@@ -84,7 +84,9 @@ public class RoleService {
     
     @Transactional
     public void delete(@NonNull Long id) {
-        roleRepo.deleteById(id);
+        if(roleRepo.findById(id).isPresent()){
+            roleRepo.deleteById(id);
+        }
     }
     
 }
